pyproject.toml
[project]
name = "imageprocessingservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.5.0", #communicates with aws s3
    "aiofiles>=25.1.0", #async file handling library
    "aiohttp>=3.13.2", #async requests library
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0", #async library for managine postgres conns
    "boto>=2.49.0",
    "fastapi[standard]>=0.121.2",
    "jwt>=1.4.0", #library for handling jwt tokens
    "pillow>=12.0.0",
    "pwdlib[argon,argon2,bcrypt]>=0.3.0", #passowrd handling library
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "s2>=0.1.9",
    "scalene>=1.5.55",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.0",
]
